LPD-65225 Don't remove the guest permissions to the portlets of private pages as we are already removing them at page level
This code was added by LPS-29597 to avoid a security problem of non-members of the site accessing to the private pages.
In Liferay 6.1 the code was executed to import both Layout and portlet permissions.
After LPS-34428 changes, it seems it is not longer used to set the Layout permissions, and only applies to portlets.
(In fact the original LPS-29597 issue was reproduced again after this change and it was resolved again in LPS-53236)