LPD-66204 Returning an instance of ServletRequestWrapper as per spec
From https://docs.oracle.com/javaee/7/api/javax/servlet/RequestDispatcher.html:

The request and response parameters must be either the same objects as were passed to the calling servlet's service method or be subclasses of the ServletRequestWrapper or ServletResponseWrapper classes that wrap them.

We need to return the original HttpServerletRequest in the getRequest
method to avoid the issue as the wrapped request must be the original
weblogic one.